+++ js/constants.js
@@ -1,0 +1,22 @@
+const MAX_ITEM = 25;
+const IMG_START = 'photos/';
+const IMG_EXTEND = '.jpg';
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+const AVATAR_START = 'img/avatar-';
+const AVATAR_EXTEND = '.svg';
+const MESSAGE_MOCK = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+const NAMES_MOCK = [
+  'Иван',
+  'Петя',
+  'Ольга',
+];
+
+export {MAX_ITEM, IMG_START, IMG_EXTEND, MIN_LIKES, MAX_LIKES, AVATAR_START, AVATAR_EXTEND, MESSAGE_MOCK, NAMES_MOCK};
